replaced taskLog logic to TaskService

+++ src/main/java/org/taskscheduler/domain/repositories/GroupRepository.java
@@ -1,8 +1,7 @@
 package org.taskscheduler.domain.repositories;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.taskscheduler.domain.entities.Group;
 
-public interface GroupRepository extends PagingAndSortingRepository<Group, Integer> {
+public interface GroupRepository extends PagingAndSortingRepository<Group, Long> {
 }
